Render in Logo-Team HTML for Tables

Appends a column with HTML that will render in team logos and names for table plotting.

Usage

cbd_gt_logos(
  data,
  team_column,
  logo_column = "team_logo",
  logo_type = "logo",
  logo_height = 25
)

Arguments

data

Pass through your plotting data

team_column

Indicate which column contains your team variable

logo_column

Indicates the name of the HTML column

logo_type

Indicate whether you want to plot logos ('logo', default) or wordmarks ('wordmark')

logo_height

The height of the logo or wordmark in the HTML string

Value

Returns data with an appended HTML column.

Details

I really hate putting team logos and names in separate columns, and this helper function will add team logos or wordmarks and create the HTML code to render in logos and teams in the same column for tables. If you are using gt to plot, you must use fmt_markdown(team_logo) to render in the HTML. Set logo_column to your team variable if you wish to overwrite your team column with the logo HTML.

Examples

if (FALSE) try(cbd_torvik_ratings(year=2023) %>% head() %>% select(team, conf) %>%
cbbdata::cbd_gt_logos(team, team) %>% gt::gt() %>% gt::fmt_markdown(team))