Generate a team matching dictionary

A utility function that returns a named vector to be used as a matching dictionary.

Usage

cbd_match_teams()

Examples

try(cbd_match_teams()['Duke Blue Devils'])
#> We notice that this is your first time logging into `cbbdata`. 
#> This process will require restarting your R session. 
#> Do you want to continue [Y/N]
#> 
#> Invalid username or password!
#> Error in cbbdata::cbd_login() : object 'resp' not found