Get NCAA Tournament Results

Returns a tibble with raw and adjusted NCAA tournament results by team, coach, conference, or seed. Data runs from 2000-present.

Usage

cbd_torvik_ncaa_results(min_year, max_year, type = "team")

Arguments

min_year

Minimum year to pull (YYYY)

max_year

Maximum year to pull (YYYY).

type

Data split value, defaults to team ('team', 'coach', 'conf', 'seed').

Value

Returns a tibble of adjusted and raw tournament results.

Details

PAKE is the number of wins above or below KenPom expectation.

Examples

try(cbd_torvik_ncaa_results(min_year=2010, max_year=2015, type='conf'))
#> # A tibble: 33 × 19
#>       rk conf   pake  pase  wins  loss w_percent   r64   r32   s16    e8    f4
#>    <dbl> <chr> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1 SEC     7.8   9.6    43    24     0.642    25    14    11    10     5
#>  2     2 Horz    6.4   7.5    10     6     0.625     6     2     2     2     2
#>  3     3 P12     6     8.4    29    20     0.592    20    16    10     3     0
#>  4     4 B10     5.4   7.5    63    37     0.63     37    28    18     9     6
#>  5     5 CAA     3     3.4     7     8     0.467     8     4     1     1     1
#>  6     6 ACC     2.8   1.3    51    30     0.63     32    24    14     7     2
#>  7     7 Amer    2.4   3.9    10     5     0.667     6     4     2     1     1
#>  8     8 MVC     1.7   2.5    12     9     0.571     9     7     3     1     1
#>  9     9 Ivy     1.4   1.8     4     6     0.4       6     3     1     0     0
#> 10    10 ASun    0.8   1.7     3     6     0.333     6     2     1     0     0
#> # ℹ 23 more rows
#> # ℹ 7 more variables: f2 <dbl>, champ <dbl>, top2 <dbl>, f4_percent <dbl>,
#> #   champ_percent <dbl>, min_year <dbl>, max_year <dbl>