Four Factor Team Splits

Split team-wide four factor data on a number of variables.

Usage

cbd_torvik_team_factors(
  year = NULL,
  venue = "all",
  game_type = "all",
  quad = "4",
  top = 0,
  start = NULL,
  end = NULL,
  no_bias = FALSE
)

Arguments

year

Chosen year

venue

Game venue ('all', 'home', 'away', 'neutral', 'road'). Defaults to 'all'.

game_type

Game type ('all', 'nc', 'conf', 'reg', 'post', 'ncaa'). Defaults to 'all'.

quad

Quad. rank of game ('1', '2', '3', '4', 'all') and quads are cumulative (games <= quad selection). Defaults to 'all', which is all games (all games Q4 or "better"/"lower").

top

Games against Top X opponents. Defaults to 0, which is "all" games.

start

Game start date (YYYYMMDD format).

end

Game end date (YYYYMMDD format).

no_bias

Logical. TRUE will display T-Rank ratings with no preseason bias/weight.

Value

Returns a tibble with four factor (and other!) data.

Examples

try(cbd_torvik_team_factors(2023, start = '20230101'))
#> We notice that this is your first time logging into `cbbdata`. 
#> This process will require restarting your R session. 
#> Do you want to continue [Y/N]
#> 
#> Invalid username or password!
#> Error in cbbdata::cbd_login() : object 'resp' not found